cleaner implementation of empty state info description
