+++ packages/client/src/components/basic/EmptyStateInfoDescription/EmptyStateInfoDescriptionStyles.ts
@@ -1,6 +1,13 @@
 import styled from "styled-components";
 
 export const StyledEmptyState = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1.5rem;
+`;
+
+export const StyledEmptyStateItem = styled.div`
   color: ${({ theme }) => theme.color["gray"][600]};
   font-size: ${({ theme }) => theme.fontSize["sm"]};
   display: flex;
