+++ packages/client/src/pages/Main/containers/AnnotatorBox/AnnotatorContainer/AnnotatorContainerStyles.ts
@@ -5,7 +5,6 @@ export const StyledEmptyStateWrapper = styled.div`
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  gap: 1.5rem;
-  padding: 0 1.5rem;
-  padding-top: 3.7rem;
+  gap: 2.2rem;
+  padding: 3.7rem 1.5rem;
 `;
