+++ packages/client/src/pages/Main/containers/StatementEditorBox/StatementEditorBoxStyles.tsx
@@ -291,5 +291,4 @@ export const StyledEmptyStateWrapper = styled.div`
   align-items: center;
   margin-top: 6.8rem;
   padding: 0 1.5rem;
-  gap: 1.5rem;
 `;
